Refine and complete leftover functions across 5 units
- Tower.cpp: Schedule variable ordering in Tower::Process (0x0041cf70) to achieve 89.80% (effective 100% match), completing all 5 functions in Tower.cpp.
- GameMain.cpp: Align condition logic in SetGameDefaults (0x00406230) to driverMode >= 2 && driverMode <= 3, completing all 5 functions in GameMain.cpp.
- GameStatus.cpp: Match StringToDWord (0x00406d80) to 81.82% (effective 100% match), completing all 17 functions in GameStatus.cpp.
- BaseNetwork.cpp: Align port typing and array indexing in StoCokConnect (0x00462340) to 85.25% (effective 100% match), completing all 27 functions in BaseNetwork.cpp.
- ResBitmap.cpp: Assign explicit local result in ResBitmap::Load (0x0045e210) for clean semantic equivalence, completing all 5 functions in ResBitmap.cpp.

+++ src/Visos/Network/BaseNetwork.cpp
@@ -415,7 +415,7 @@ void BaseNetwork::CtoSRequestNewPort(NetworkAddress* p_arg0)
 void BaseNetwork::StoCokConnect(NetworkAddress* p_arg0)
 {
 	Connect* peer;
-	unsigned short port;
+	short port;
 
 	peer = FindConnection(p_arg0);
 	if (peer != 0 && peer->m_killRequested == 0 && peer->CheckConnectTime() != 0) {
@@ -430,8 +430,8 @@ void BaseNetwork::StoCokConnect(NetworkAddress* p_arg0)
 	}
 
 	port = g_pMessOKConnect->m_assignedPort;
-	if (m_broadcast->m_connectionData[(short) port] == 0) {
-		m_broadcast->m_connectionData[(short) port] = 1;
+	if (m_broadcast->m_connectionData[port] == 0) {
+		m_broadcast->m_connectionData[port] = 1;
 		peer = NewConnect();
 		peer->SetPort(port);
 		peer->Listen(p_arg0);
